backend/requirements.txt
Flask
flask-cors
openai
azure-storage-blob
azure-cognitiveservices-speech
python-dotenv
fpdf
numpy
faiss-cpu
requests
dotenv